+++ containers/templatetags/container_tags.py
@@ -7,6 +7,11 @@ from containers.models import (
     STATE_RUNNING,
     STATE_EXITED,
     STATE_INITIAL,
+    STATE_PULLING,
+    ACTION_STOP,
+    ACTION_START,
+    ACTION_UNPAUSE,
+    ACTION_RESTART,
 )
 
 
@@ -22,6 +27,22 @@ def colorize_state(state):
         STATE_EXITED: "text-secondary",
     }
     return colormap.get(state, "text-dark")
+
+
+@register.simple_tag
+def state_bell(state, last_action):
+    if last_action is None:
+        return ""
+
+    if state in (STATE_RUNNING, STATE_PULLING):
+        if last_action not in (ACTION_START, ACTION_UNPAUSE, ACTION_RESTART):
+            return "Should be running or pulling"
+
+    elif state is STATE_EXITED:
+        if last_action is not ACTION_STOP:
+            return "Should be exited"
+
+    return ""
 
 
 @register.filter