+++ containers/tests/test_templatetags.py
@@ -8,8 +8,16 @@ from containers.models import (
     STATE_RUNNING,
     STATE_FAILED,
     STATE_EXITED,
+    ACTION_START,
+    ACTION_STOP,
+    ACTION_PAUSE,
+    STATE_PULLING,
 )
-from containers.templatetags.container_tags import colorize_state, pretty_json
+from containers.templatetags.container_tags import (
+    colorize_state,
+    pretty_json,
+    state_bell,
+)
 
 
 class TestContainerTags(TestCase):
@@ -51,3 +59,42 @@ class TestContainerTags(TestCase):
     "key4": "value4"
 }"""
         self.assertEqual(pretty_json(json.loads(data)), expected)
+
+    def test_state_bell_no_action(self):
+        self.assertEqual(state_bell(STATE_FAILED, None), "")
+
+    def test_state_bell_state_running_action_consistent(self):
+        self.assertEqual(state_bell(STATE_RUNNING, ACTION_START), "")
+
+    def test_state_bell_state_running_action_stop(self):
+        self.assertEqual(
+            state_bell(STATE_RUNNING, ACTION_STOP),
+            "Should be running or pulling",
+        )
+
+    def test_state_bell_state_running_action_pause(self):
+        self.assertEqual(
+            state_bell(STATE_RUNNING, ACTION_PAUSE),
+            "Should be running or pulling",
+        )
+
+    def test_state_bell_state_pulling_action_stop(self):
+        self.assertEqual(
+            state_bell(STATE_PULLING, ACTION_STOP),
+            "Should be running or pulling",
+        )
+
+    def test_state_bell_state_pulling_action_pause(self):
+        self.assertEqual(
+            state_bell(STATE_PULLING, ACTION_PAUSE),
+            "Should be running or pulling",
+        )
+
+    def test_state_bell_state_exited_action_consistent(self):
+        self.assertEqual(state_bell(STATE_EXITED, ACTION_STOP), "")
+
+    def test_state_bell_state_exited_action_start(self):
+        self.assertEqual(
+            state_bell(STATE_EXITED, ACTION_START),
+            "Should be exited",
+        )